e2e: update e2e test
